requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-shopping-cart==0.1
gunicorn==20.1.0
idna==3.2
phonenumbers==8.12.28
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
